Refactor director stage window to abstUI layout
